Vendor JSON5 package and emit unquoted object keys if possible
The original JSON5 package from github user yosuke-furukawa was the Go stdlib json package with modifications but
it was missing the necessary license information and did not provide any license information.  As it did not
encode simple keys unquoted I had to modify it anyway.  Put it under third_party to make it clear it not originaly
my code.