Fix color histograms with dominant colors
If an image has one dominant color (e.g. green in the Unic logo) then the other colors
will have very low counts. If these counts are not properly rounded such images will
result in string representation of the hash which has just one 'f' because all colors
with a hash count < 16 are rounded down and will be printed es zero.  These effects can
accumulated, e.g. if 8 colors with a count of 14 are dropped this rounding while generating
a string representation will represent up to 40% of the maximum count which is non-neglible.
The effect was noticed because the actual Unic logo image had an strange ColorHistDelta
to the calculated fingerprint.