Merge package check and package response into ht
Seperation of checks and general testing code still would be desirable,
unfortunately it results in complicated code and makes some things almost
undoable. E.g. the W3CValidHTML check needs to craft a HTTP POST request
and evaluate the response which would be very easy if this check could use
a ht.Test which it cannot because this would result in a circular reference.
Breaking the circular reference between ht and check with the extra
package response works but is ugly and dosn't solve the problem of the Link
check which should check links in the same context/session as the original
test (which again would result in a circular reference).
All this indicates that the coupling of packeges check and ht is much
tighter than two seperate packages suggest. So unify them in the hope
that this will also allow to simplify the result handling of a test
(which is currently not really nice).