Keep test results in the test itself
Instead of returning a special type TestResult which contains lots of
data copied from the original test store the outcome of a test execution
directly in the test. This is a bit awkward if tests are reused (e.g.
during load and performance testing) but it should result in a much
saner design and less copying.
The intended 'fix' for the load and perfromance measurements is to add
special (small) types to capture the relevant outcome of a performance
measurement, i.e. duration and status and to add an option to log
failing tests/checks: A normal performance measurement should not result
in 20% failing tests and drilldowns into failed test of a load tests
seem overly fancy.