Improve test specification
The old way was an obscure mixture where the JSON5 representation of a Suite
could contain Tests, but Tests could live in other suites, other files or
even on their own and all Tests and Mixins where referenced by their Name field.
This was horrible.

Now the JSON5 spec of a Suite contains basically just some parameters for the suite
and references all Tests via their filename. So each Test and each Mixin now lives
in its own file and Suite reference tests and Tests reference Mixins via the filename.
Instead of one long JSON5 document describing a suite _and_ the tests there are now
several small (and shallower) JSON5 documents which makes handling easier.

For now the tests and mixins must be in the same directory as the suite, but that
could be relaxed by allowing relative filenames.

During simplifying the JSON5 documents for a suite the Variable section was
included into the Suite itself. This is a bit illogical but the shorter and
simpler definition justifies it.