Fix panic from reading a nil body
The error condition triggers if the response body is empty
but the content encoding isset to gzip: Under this condition
the error handling was lacking and resulted in reading from
a nil body.

The current fix is a stopgap. Maybe the right way to handle this
would bo to not report an error instead of 'gzip EOF'.