fingerprint: implement image reconstruction from BMV hash
For debuging purpose it is convenient to recunstruct the original image from the
fingerprint hashes.  Of course this yields only a very raw reconstruction, e.g.
the BMV hash has only a few bits of information at all.

The generated image is not completely black/white but verydark/lightgray
simply because this is much nicer to the human eye.