ht: add loading of single tests
Loading of single Tests from a JSON5 files is added as it is clumsy to pack
tests into suites just for running them.

During this a bug and an inconsitency was found in the loading code:
To make accessing stuff homogenious the new code now uses the relative
filename to find mixins specified in the BasedOn field like a suite
does for tests.