Rename the sample suite
Now that filenames are used to reference tests and mixins inside a
suite it seems natural to rename the example suite 'suite.ht' to
'sample.suite' which is clearer: Its extension indicates a suite
and the base name indicates it is an example.

To be consistent: Reword the command helps for cmd/ht.