cmd/ht: use normal exec code during 'run' subcommand
The 'run' subcommand creates an arteficial suite to run the individual test
files, so there is no need to double the code for suite execution and result
printing in run: Just hand over the autogenerated suite to the exec functions.

While at it: Make 'run' understand the -D command line flag. Without the
ability to set variables it is often impossible to run a single file test.