ht: fix reporting errored test in JUnit XML report
If the requests fails (e.g. because of a timeout) the test gets status 'error' and
the checks are skipped. Naively reporting every check as a JUnit testcase will
report any such not run check as skipped. This is missleading

Thus report such as errors. This double-counts the test error for each skipped
check but the alternatives look even 'more wrong': Dropping the checks changes
the total number of testcases. Maybe reporting just one testcase as error and the
other as skipped would be sensible.