ht: fix skipping checks after failing StatusCode==200
Both StatusCode and *StatusCode satisfy the Check interface. When deserializing a test from JSON5
the status code checks are of type *StatusCode. This results in all checks beeing executed, even
if the first check is a failing status code check against 200 because the type assertion is
against StatusCode while the actual type is *StatusCode and the extra skipping logic was not
triggered.

This bug went undetected for a long time as all the tests are typically written with StatusCode
and not *StatusCode.

Fix it by type asserting the first check twice and a largish test covering both, StatusCode and
*StatusCode.