cmd/ht: report proper exit codes
Exiting always with 0 after running a suite (no matter how many errors or
failures) is a clear bug. Thus report bogus as 3, error as 2 and just check
failures as 1 and passing everything as 0.

While at it consolidate the exit codes in the help command and ht itself:
Report a wrong invocation of ht or an unknown command with an exit code of 9.
A valide 'ht help list' exits with 0.