ht: implement log file checking
This commit add some basic functionalities to check log files: The data written to a file
between preparation of the check (i.e. before the request) and executing the check (i.e. after
the request) can be checkd for text content and size.
This is intrinsically unsafe in systems where other requests may happen or background tasks
may produce additional log messages; but it is better than nothing.