all: rename HTMLContains and HTMLContainsText
HTMLContains was looking for specific tags and HTMLContainsText was checking
the text content of specific tags. These names were long and not very clear.
Thus make the following renaming:
    HTMLContains      -->  HTMLTag
    HTMLContainsText  -->  HTMLContains
This should be much clearer. E.g. HTMLTag looks for tags or a certain number
of tags and HTMLContains checks for text in the HTML document.

Add examples to the sample tests.