cmd/ht: add flag to display checks
It is hard to debug tests which are read from a JSON5 representation:
If you make small mistake, e.g. a typo like 'Slektor' instead of 'Selector'
in a check this will be unmarshaled without error but the generated check
is broken as the Selector field is zero.
Also the mixing of tests might add checks. All this makes it diffucult to
clearly see what is going on.

Thus add a -check flag to the list subcommand which prints the checks.
This allows to recognise typos (if the printed check differs from what your
JSON5 contains) and to see clearly what is going on.