ht: fix loading of nonlocal tests
If all tests (and all used mixins) of a suite are kept in just one folder
everything wored fine. But using non-local tests with non-local mixins was
broken: A local mixin from a non-local test was not found.

Now if a suite references a test as 'some/path/test.ht' and this tests
is based on a mixin ../foo/m.mix then m.mix will be read from 'some/foo/m.mix'.
Add some testdata to test the desired behaviour.
While fixing this non-local import probelm it was discovered that merging of
request headers and parameters was not a real merge: fix this too.

The logic is still broken if multiple suites from different folders are
loaded.