ht: add business criticality field to test and allow fine grained reporting
Up to now tests have been 'all or nothing', pass or fail. This is annoying if used in
larger suites where one might want to include tests which are interesting but not of
high business criticality or importance.  The new Criticality field is just a label which
is taken along the lifetime of the different tests.
To faciliate reporting over one or more suites a new type SuiteResult is added. This
type allows to accumulate suites once executed and produce KPIs based on user defined
panalty functions. Hopefuly these KPI will be valuable in monitoring web applications
as such a KPI condeses several suites into one number which can be monitored much
more easily than the complex outcom of a suite execution.