ht: add exploded JSON condition testing
The gojee expressions are not realy native: Their syntax is strange and
a mini-language by themself which makes writing such checks complicated.

Thus add checks based on the standard Condition type: A JSON value is selected
by it's path down to the element. This flatteing is done via gojsonexplode
which works qute nice.