ht: make Extractor an interface and add BodyExtractor
Extraction of HTML elements wont be sufficient in the long run. Thus make Extractor and interface
move the HTML extraction to HTMLExtractor and add a new BodyExtractor which does regexp based
extraction from an uninterpreted body.

For (un)marshaling of the extractor map of a test ad a type ExtractorMap which works
like the CheckList and allows attaching the JSON5 (un)marshaling code. Extractor
registration works like the one for Checks. The NameOf function is even identical
thus unify it.