ht: process special variables in defined order
To get reproducable random numbers and text it is necessary to determine the values
of random variables in a defined order. Storing the random (and now) variables in a
map yields them in arbitrary, unfixed order. Thus keep them in a sorted slice.
Add a real test for the variable asignement code.