ht: fix dropping of FollowRedirect policy for loaded tests
If a test is loaded from JSON5 and is based on an other test
the FollowRedirects policy of the request was not properly propagated
to the resulting test. Maybe this is a bug of the merge code but
for now fix it in the loading from JSON5 code by direct setting
the policy to what was requested in the JSON5 test.

(Merging of tests uses the strategy 'last wins' for the follow redirect
policy which probably would work if the order in which the tests are
merged during loading would be 'base test first, then actual test'.
So the bugfix is a bit ugly and hacky as it does not address the
fundamental missmatch between how Merge works and how Merge is used
in the loading from JSON5 code.)