ht: get rid of external package for pretty printing
The test code imported github.com/kr/pretty to pretty printsome stuff,
mostly during reporting errors.  This package is a dependency which
makes ht not selfcontained. The other needed dependencies are either
internal or vendored.
Reworking the few uses of package pretty is straightforward, so let's
get rid of this dependency.