ht: add explicit DeleteCookie check
The SetCookie check can check for a deletion, but the whole idea of
checks was to provide human readable high-level checks. The SetCookie
check with a 'Delete bool' field is not human readable (or easy to write),
it is a monster.
Thus seperate the check for a cookie deletion into an own DeleteCookie
check which is much simpler and more robust to use.