ht: record and report redirections
Not following redirects, checking the Location header and doing the next request
manualy is okay sometimes. Often it is not as it is too cumbersome; that is the reason
why FollowRedirects exists: Make a request and see the actual outcome.
Unfortunately this automatic following a redirect chain is completele opaque to the
user, only the final URL requested is recorded. This makes debugging such request
chain difficult to impossible.

The solution implemented in this change records the URLs of all intermediate and the
URL of the last request made. This redirections record is available to be inspected
by checks and is reported in the text and HTML output.

The main drawback of this solution is, that each test which automatically follows
redirects needs its own http.Client where the CheckRedirect function is a closure
and records the redirects in the test. This means that follow-redirect-tests cannot
reuse a client from the client pool. But maybe this whole pool of clients is a
bad idea anyway, or at least unnecessary for checking responses (and professional
loadtesting was never a real goal).