ht: add Redirect check
Checking redirections is possible with the Header check, but it is clumsy, e.g.
it requires two checks, one for a StatusCode of 302 and one for the expected
Location header.

The new check Redirect combines StatusCode test and Location header test in
one check with a pleasant to read syntax.

The three dots '...' to indicate prefix/suffix matching is uncommon in ht,
normaly a Condition would be used. But Conditions are a lot to type and most
of the functionality of a Condition is probably not needed to check the
Location header. But checking that the rederections goes to https or to
a certain path is very common and very readable with the syntax
    To: "https://..."   or   To: ".../some/page.html"