ht: improve semantics of HTMLContains check
The current code required that the expected Text field contains
the first actual matches found in the HTML and these in the found
order. This makes testing complicated as it is hard to express
the intention "i'd expect this item to be present in that HTML
slot but there may be others too".

Change the semantics so that the check makes sure all expected
entries in Text are present in the HTML. If order matters a new
field InOrder makes sure the expected text are found in the given
order. To check for identity you set InOrder and Complete to
true.

This newsemantics does not allow to check smething like "the
first 3 elements are ...", but this seems to be less common.