cmd/ht: seed random number generator of ht
Some applications need unique values each time they are called, especially
if they do have state and are not properly reset after running the test.
This is a problem with some legacy software systems in the enterprise
segment. The problem can be overcome with ht by using e.g. a {{RANDOM NUMBER}}.
But to be rally different on each run we need a different seed.

cmd/ht now seeds the random number generator used in ht with the current
timestamp. To allow reproducible reruns of a suite the seed is exposed
on the command line and can be set determiistically.