ht: get rid of client pool
While adding the code to record the redirect chain individual clients are
constructed, but these do not have a cookie jar which leads to cookies
not beeing sent on redirected requests. This is bad and was always present
if no client pool was used. As most complex tests where executed via a
suite this was not obvious as the suite alwys provides a client pool.

Get rid of this client pool abomination and expose the Jar individualy
in the Test instead of hiding it in a client pool. Maybe the Transport
should be exposed in a similiar fashion.