cmd/ht: add reading variables from a file
The variable substitution in ht has basically two uses: a) during unrolling a test which
gives kinda table-driven testing and b) to parametrize a tests to a certain environment
which allows to set e.g. hostnames from outside of the test.

For the later use the -D command line flags to cmd/ht are useful but setting more than one
or two like this becomes clumsy; it feels a bit like using curl(1) which was not the
intention of ht.
Suites allready allow to define a Variables section which is quite nice but running
a suite with different variables requires lots of '-D var=val' on the command line.

This changes allow to read a bunch of variable=value pairs from a JSON5 file with the
new '-Dfile <file.json>' command line flag.