Fix overwriting the cookiejars when preparing suites
The suite would overwrite any cookiejar set on teh test during the prepare phase.
This is okay for direct test read from JSON files. But it prevents keeping
cookies during checking links: The generated test should use the current
cookiejar but that setting is overwritten with a new and empty jar during
prepaation of the suite.
This change just prevents overwriting any pres-set cookiejar which probably
is the most sensible way to handle this.