ht: add new check type Latency
The new Latency check redoes the original request several times and checks
that sample quantiles of the response times do not exced the given limits.
This should allow to do simple performance tests right during functional
testing.
The handling of check in the redone tests is a bit coarse: You can disable
the tests completely or run them all but the Latency tests (to avoid an
infinite recursion).

The code has one major issue -- variable handling: The latency-measuring
requests are done with a copy of the original test ad nothing special
is done for the variables. Fixing this might require major rework of
how tests are executed.