ht: add Sorted check
Checking that something is returend in sorted order is very common.
Such a check can be made with a simple regular expresion in the existing
Body check but such a check doesn't properly convey that you want to
check for the proper order of something.
This is where the new check Sorted shines: It clearly conveys 'we check
the order of the given elements'.
The chekc is currently not so suitable for HTML output.