ht: improve Sorted check for HTML pages
Applying the Sorted check to HTML pages is possible but might be brittle:
Assume you want to check some list is sorted by price so you do a
    Sorted{Text: [ "CHF 12.00", "CHF 45.95", "CHF 107.70", "CHF 834.55" ]}
which works fine as long as "CHF 12.00" is literaly present in the HTML
page but will fail if the HTML is styled like e.g. in
    <p>CHF <span class="cheap">12.00</span><p>

This commit add automatic text extraction and normalization to HTML pages,
so that the Sorted check would "see" just the plain text "CHF 12.00", even
in the styled output. THis should make checking HTML pages much more robust.