ht: limit number of missing values in Sorted check
The boolean AllowMissing is too coarse grained: E.g. in a list of 10 elements
three missing might be okay, but three missing in a list of 5 element (or even 4)
renders the test nonsensical.
Thus limit the toal number of allowed elements to be missed and report all
misses during failure.