json5: add a Marshaler5 to produce JSON5 for custom marshaled stuff
Marshaling the check list looked nice but produced entries like
    { "Check": "ContentType", Is: "html" }
with unnecesarry quotes around Check.  This could have been "fixed"
in the MarshalJSON method by just not printing these quotes but that seems
wrong: The output would not be well-formed JSON.

Thus add a new interface Marshaler5 to the json5 package which can be
implemented by types who which to be marshaled as JSON5. CheckList has this
whish.