recorder: improve handling of POST parameters
POST requests may contain parameters in the body. Keeping them
in encoded form in the body is possible but errorprone and basically
unreadable. Thus scan the POST body for parameters and record them
in the Params (and the encoding method in ParamsAs) field of the
generated test.

While at it:
 - Cleanup logging.
 - Do not produce content based checks for redirect responses.