ht: add NoServerError check
Sometimes all you care is that the server doesn't throw errors, e.g. one
might not care wheter teh server returns 200 or 302 or even 404 because all
you want to check is that the server behaves properly.

This is especially useful in automatically generated test, e.g. in the upcomming
reslience check.