ht: new check Resilience
Web applications should work properly even if the requests made to the
application is not well formed or in the desired form. Applications
use the HTTP methods to detect actions, HTTP and header fields (e.g. to
detect language or browsers) and parameters to determine their behaviour.
All these are user-controlled and a malicious user (or a broken client)
may send unexpected or even malformed data.

The new check Resilience determines the basic resilence of a certain
URL in the application against unexpected and malformed input.
It checks a large amount of manipulations to the original request
made and fails if the server returns any 5xx status code.

With this it was easy to find the following two problems in two
applications:
 - Insufficient validation of a query parameter in a search.
 - A server error on a POST request with malformed User-Agent string.