ht: keep cookies during resilience checks
Without the cookies the checks might behave differently.
This is just a stopgap solution: The requests made during resilience
checking might modify the cookiejar which is undesirable.
The proper solution would be to manually copy the cookies
from the original jar and send together with the request.

While at it: Surround the resilience suite with log messages
to make the logfile a bit more readable.