ht: use static version of cookies in resilience check
It is important to send the appropriate cookies in the resilience check
requests, but it is probably a bad idea to use the original cookie jar
from the original test as the server might modify cookies.
Thus extract the cookies whihc would be sent from the jar and add them
to the resilience-requests statically.