ht: stopgap fix for missing bodies in redirect responses
A 3xx response may contain a body but this is not available if the
request was done via a http.Client which is configured to not follow
redircts. (This can be done by using a http.RoundTripper).
The old code tried to read from the body which had been closed by the
http.Client, resulting in an error.
The new code just ignores the body of any 3xx type response. This is
still broken, but arguably less broken than the old behaviour: Typically
you either follow the redirect chain to read the final body (and check
the body) or you do not follow the redirects and inspect the Location
header or check that cookies are set or deleted properly (and ignore
the body).