json5: encode channels as null
It is handy to encode e.g. a http.Request or a http.Response as JSON5 as it
works like a pretty printer. Unfortunately channels are un-encodable (of
course). As we missuse the JSON5 package anyway: Encode chan types too,
allways as "null". This might prevent roundtripping but we do not care.