ht: change type of read body
The body read from the response was traditionally stored as a []byte.
Often this []byte had to be converted to a string, there is no real
benefit from a byte slice especially as the body is considered to be
immutable once it has been read.
A byte slice has one major drawback: It serialises via base64 encoding
during JSON marshaling which renders the body unreadable which is
impractical, especially as we gunzip it before storing it as the main
point is having a directly readable body ready at hand.

Thus change the type of the stored body to string.