ht: add debug output to resilience check
Finding a reslilience problem is easy with the resilience check, but the
reported error is not very verbose and usable (albeit technically complete).
This CL allows to save the failed resilience tests to a file, including
everything, from the sent http.Request to the received http.Response
including the whole body. Everything almost ready to be re-executed.

This should make debugging or writing sepcialised tests for an individual
problem much easier.