ht: make ValidHTML useable
Parse the HTML document, record tag nesting structure and
check for multiple ids or missing ids and doubled attributes.
Checking of proper encoding is implemented very rudimentary.
Handling of URLs is prepared but currently broken.

Several checks produce not just one, but a list of errors. Handling them
as strings is painful, thus introduce a special error type which is a
whole list of errors.
While at it: Errors with position information are quite useful.