ht: fix validation of special HTML elements
The content of some HTML elements (script, noscript and iframe) is not parsed by
package html and cannot be check for broken escaping as it will allways fail.
Omit these elements in the escaping check. Should be save to do as these are
rare.

Some elements are empty and can be written as selfclosing or just empty,
both is valid HTML5. So do not start a new block if such an start tag is
found.