ht: structure output of multi-error checks
Several checks may return multiple errors, e.g. multiple issues found during
HTML validation, or multiple latency limits which are exceeded. Honour this fact
an output all the errors individually in the generated text output.

Still to do: Use this in the HTML output and maybe even in the JUnit-output.