cmd/ht: fix error handling in fingerprinting subcommand
If an image fingerprint of a URL was requested and the GET to this URL returned
e.g. 404 or 401 this error was not reported but the body was decoded which
then failed (as as typical 404 body is not an image) and the displayed error
('unrecocnized image format') was completely unrelated to the root cause.

Return the HTTP status if anything other than a 200 status code is received.