cmd/ht: add new subcommand 'debug'
Writing and debugging tests and checks should be as easy as possible. The
current solution requires to save the test/check to a file (in JSON5 format).
Execution is done via the command line and the full report is written to a
HTML file. Debugging requires consant context switches which are painful.

Thus allow to run a test directly from the command line by passing the JSON5
as a command line argument.  This won't work for full fletched tests, but
to debug a single check it is okay and very helpful.