ht: include 'home' or 'up' link into generated HTML report
While experimentation with tests and suites one typically generates lots
of folders with the results. Navigating from one report the a new one is
a bit cumbersome as you have to delete parts of the URL in the browser
address line.

With this CL a relative link to '../../' is displayer automatically
which makes browsing to the newest report much eaiser.