recorder: rewrite absolute URLs and fix gziped content
Sometimes the response contains absolute URLS, especially in redirects
or as absolute links in a- or img-tags. This makes capturing events
difficult as one must manually change back the origin hostname to
localhost:8080 to hit the recording reverse proxy.
To faciliate this: Rewrite Location headers and href and src attributes
to localhost. This is done in 'brute force' and might replace too much,
especially any literal occurence of 'href="http://origin.host.org' but
this should be uncommon.

Found while doing the above: If the server sends gziped content we have
to unzip it first before a) trying the above rewrite of absolute URLs
and b) try to extract HTML checks.
So unzip (and rezip) the content if needed.