ht: implement inverted checks
This is kind a hack, but a nice one: Several checks already allow to reverse
the condition checked, e.g. ResponseTime has Lower and Higher fields and
any Condition can be reversed by setting a negative Count. Sometimes this
works pretty well (ResponseTime), sometimes it can be done but it feels
a bit awkward (Condition) and sometimes the check cannot be reverted (e.g.
UTF8Encoded) at all.

This CL introduces the notion of a negated check: In the JSON5 serialisation
it is spelled by prefixing the check name with "NOT ", e.g.
    {Check: "NOT StatusCode", Expect: 404}
which seemd like the most readable way even if ugly.

The intended use case for this change is twofold:

 1. To increase readability/writeability for Condition based checks, e.g.
        {Check: "NOT Body", Contains: "badvalue"}
    instead of the traditional
        {Check: "Body", Contains: "badvalue", Count: -1}

 2. To make some strange checks expressible in a natural way, e.g.
        {Check: "NOT Body", Prefix: "Failed"}
    instead of some weird regexp dance.

Especially point 2 might be useful once we allow specialised checks inside
some checks, e.g. Resilience just checks for NoServerError which is crude.