ht: add logical OR if checks and refactor logical NOT
While adding the logical OR of checks it got obvious that the clever way of
negating a check was overly clever and too hacky. This CL implements
boolean OR and NOR; both are short circuiting. The implementation and
how to specify these gates is homogenious and doesn't use much clever
tricks.
The Go types have a bit strange names, e.g. AnyOne with its only field Of
but this tries to mimick natural language: 'Check any one of the checks ...
is true' and 'Check none of the checks ... is true'.

Unittesting is subpar.