ht: add new check RedirectChain
Checking a chain of redirections can be done with several tests, each
not following the redirect and each with a Redirect check to check the
Location header and the StatusCode. But this is a lot of work and brittle
as the chain is distributed on sevaral tests. A check for the final URL
exists but cannot check the intermediate steps in a longer redirection
chain.

Ths CL implements a new check RedirectChain which allows to check some
or all steps in a redirect chain. Of course the check is pretty coars,
e.g. you cannot check the StatusCode and you cannot check that cookies
are properly set during the chain. But it should be a good pre-test
to make sure the chain itself is properly traversed (without loops)
and subsequent tests may check details of individual steps.