ht: improve error reported for bad checks
If a check cannot be unmarshaled rom JSON5 because a field was misspelled
the error reported was not very helpful as this unmarshaling happened not inside
a 'real' file, but inside the check itself as each check itself is unmarshaled
from a reencoded byte slice.

Improve the reported error a bit by displaying which check failed instead of
the always constant line number 1.