fingerprint: keep more seldomly used colors in color histogram
When fingerprinting logos or product images which contain a lot of homogenious
background and just a few dominant colors in the foreground it happens that
several seldomly (i.e. not dominant) colors get 'dropped' while converting the
[24]byte histogram to a 24 hex digit string: As each hex digit has just 4 bits
any color which occurs less than 6% compared to the most dominant color is
completely lost as it will be printed as 0.

This problem does not occur for 'natural' images as highly dominant colors
are less prominent in such images. The problem was discovered while fingerprinting
a logo with a transparent background (which is treated as black).

To mitigate this problem a bit the histogram undergoes a nonlinear transform
which includes more seldomly occuring colors in the string fingerprint.
The relevant constant gamma in this nonlinear function was choosen after
some manual experimentation with two different logos.