fingerprint: use 5 instead if 4 bits in color fingerprint
The color histogram was encoded as a hex string (4 bits per character)
which is nice and consistent with the BMV fingerprint but it is very
coars. That's why some fancy nonlinear transformation had to be implemented
to keep colors from beeing 'dominated out of the fingerprint'.

But actually we never use the fact that the fingerprint is a hex number;
it is just a fingerprint.
Thus encode the histogram in 5-bit values i.g. the characters 0-9 and a-v.
This also alloed to reduce the nonlinearity (gamma).