ht: fix SetCookie check
Checking the domain against the value of a cookie cannot work. Fix it.

While at it: The error reported when the cookie Value or the Path or
the Domain are not okay are very terse, e.g. just 'NotFound' if the
regexp doesn't match. This can be missleading: The cookie itself was
found, just some subfield didn't match. Wrap the terse error into
something more readable.