ht: add screenshot checks via PhantomJS
For our purpose it doesn't seem necessary to keep a single, long running
PhantomJS instance around as sharing such an instance between checks or
even tests is not really possible. So let's just invoke PhantomJS directly
without using a wrapper. Instead collect all cod into a temporary script
and execute this script.

Difference checking is pretty crude. Any visually different pixel is
counted an there is a hard tolerance limit on the number of different
pixels. The idea of preprocessing the difference image, e.g. by supressing
isolated "noise" differences might be useless as these are generated,
rendered images and not photographs. If there is a difference, even a
single pixel then this is not "noise" but a difference as the rendering
process is deterministic.

TODO: The color difference uses a third-party library to compute Delta E.
Again this might be overkill because of the same reason as above: Any
difference is a difference, even if invisible to the human eye.

TODO: It might be useful to exclude regions of the screenshot from the
comparison. Currently this can be done through visualy disabeleing the
element via JavaScript, but this is error prone and complicated and might
not be possible in all cases without side effects.