ht: make PhantomJS executable a global variable
Keeping the PhantomJS executable (path) in a field of the Screenshot tests works
but is pretty strange as nobody would want to use different versions of PhantomJS
for different screenshots. At least nobody sane.
Thus move this setting to a global variable which is exported and thus settable
for users of package ht.