ht: make file a test was loaded from available in variables
Screenshot checking requieres a 'golden record' of the expected screenshot. Up to now
it was not possible to state a path to such a golden record relative to the file
the test was read, only absolute paths worked, or relative paths to the current
working directory.  This is ugly.
Add variables to each test and populate the tree variables TEST_NAME, TEST_DIR and
TEST_PATH with the basename, dirname and absolute full path of the file the test
was read from.