cmd/ht: use a cookiejar even when 'run'ing a single test
If a single test is execeuted via 'ht run <test>' an arteficial suite is generated.
Up to now this suite did not keep cookies and thus the test did not have a cookiejar.
This seems totaly sensible as just one request is made by the test. Unfortunately
this is no longer true as several checks generate some or even an awful lot of
secundary requests. FOr several of them it would clearly be nice to have the
cookies received from the original request available; think of the Links checks.

Of course the user could handcraft a suite around this singular test and execute
the suite, but this is not the intended purpose of 'go run'.

So just keep cookies when running a single test via the ht command.