cookiejar: new package
We are using the stdlib cookiejar to keep cookies between requests which works fine
but has one drawback: The stdlib jar is completely opaque and it is almost impossible
to get the cookies of of the jar.
Until now this has not been a problem as the jar is used by the HTTP clients and each
request goes through this client (or several client share the same jar).

With the new PhantomJS based checks it would be nice if we could populate the
jar of PhantomJS with all the cookies in our clients jar.

The new package is a fork of the stdlib net/http/cookiejar with one main change:
Its content is inspectable. The fork also provides notification on changes but this
feature is not used here.