ht: allow reading request body from file
Sending some JSON as the request body is awful if done through cmd/ht:
The test is JSON5 and if one wnats to include an other JSON as the body
string you have to quote all quotes (and newlines).

Experimentaly allowing the body to be read from a file like parameters.