ht: add tests for rendered/processed HTML
Rendered/processed in the sense of: A headless browser (PhantomJS) has parsed the
raw HTML and loaded the linked scripts (and images, and CSS, whatever) and processed
the JavaScript found (which typically modifies the raw HTML to yield the final
DOM to be displayed.
Defining such checks is okay, unfortunately these checks are kinda 'sub-checks'
and reporting is not nice and error reporting for failed checks is ugly and
the rendered HTML is not part of the test report at all.