ht: improve file upload for parameters and body
Parameters and the request body treated values starting with "@file:"
special, the read the value from the file given after this prefix.
This worked in simple cases but e.g. the upload of several files
with the same parameter name was not possible (only the first was
processed), the code was ugly and test coverage low.
Additionaly it was not possible to substitute values into the
file content which seems to be a sensible idea.

This CL adds a new form of the @file: values, the @vfile: values
which work like the @file: ones but do perform variable substitution
on the file content. This works for parameters as well as for the
request body.

While doing this: Refactor the code, fix the known issues and
increase the test coverage.

Update the tutorial and improve it by explaining NOW and RANDOM
variables.