ht: add simple Basic Auth to requests
Basic Auth is common, currently the Authentication header has to be crafted manually
which is dead ugly when executing test stored as .ht files on disk as you have to
base64 encode username and password.

Add two more fields to the request: BasicAuthUser and BasicAuthPass. The fields are
a bit long to type but you can copy/paste it or even extract them to a mixin
and include it via BasedOn.