all: move timeout to request
Up to now the timeout was a field of Test but it never was a timeout to the
Test, but only to the Request. This is stupid, thus move it to the Request
where it belongs.

This change is not backward compatible and will break all timout related
code and test. Better now than never.