all: remove 'business criticality'
This was a bad idea. Totally stupid. A test or a check tests desired behaviour.
If this desired behaviour is not relevant or has a low 'business criticality'
it is either not really desired or not worth testing/checking.

The criticality was used only for the monitor subcommand of cmd/ht. That it
was a bad idea could be seen at the different penealty functions: It simply was
not obvious how to rate a moderately critical test which failed against one
which errored.

This seemd as a good idea but it is not. Removing it deleted quite some
code which is always nice.

This is a change which breaks backward compatibility. Do it now.